+++ user-service/src/main/java/com/palja/user_service/application/event/dto/impl/DeleteCompanyUserEvent.java
@@ -1,6 +1,8 @@
-package com.palja.user_service.application.event.dto;
+package com.palja.user_service.application.event.dto.impl;
 
 import java.util.UUID;
+
+import com.palja.user_service.application.event.dto.UserEvent;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
