+++ user-service/src/main/java/com/palja/user_service/application/service/impl/CustomerServiceImpl.java
@@ -18,7 +18,7 @@ import com.palja.user_service.application.dto.response.CreateUserRes;
 import com.palja.user_service.application.dto.response.ReadCustomerDetailRes;
 import com.palja.user_service.application.dto.response.ReadCustomerSummaryRes;
 import com.palja.user_service.application.dto.response.UpdateCustomerDetailRes;
-import com.palja.user_service.application.event.dto.DeleteCustomerEvent;
+import com.palja.user_service.application.event.dto.impl.DeleteCustomerEvent;
 import com.palja.user_service.application.exception.AuthErrorCode;
 import com.palja.user_service.application.exception.UserErrorCode;
 import com.palja.user_service.application.service.CustomerService;
